Use C++ math functions overloads instead of C float versions in std
libstdc++ does not provide overloads like std::floorf.
A bug due to a macro made this compile accidentally before.

+++ src/modm/ui/animation/interpolation.hpp
@@ -75,7 +75,7 @@ private:
 			float delta = (static_cast<float>(end) - begin);
 			deltaValue = delta / steps;
 			if (deltaValue == 0)
-				deltaValue = std::copysignf(std::numeric_limits<float>::epsilon(), delta);
+				deltaValue = std::copysign(std::numeric_limits<float>::epsilon(), delta);
 			accumulatedValue = static_cast<float>(begin) + deltaValue / 2;
 		}
